requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
watchdog>=3.0.0
Flask>=2.3.0

# PDF processing libraries
pdfminer.six>=20221105
PyPDF2>=3.0.0
pymupdf>=1.23.0
pdfplumber>=0.10.0
pdfrw==0.4
reportlab>=4.0.0
Pillow>=10.0.0

# Word document processing libraries  
python-docx>=1.0.0
docx2txt>=0.8
docxtpl>=0.16.0
Jinja2>=3.0.0

# Optional: Comprehensive document processing (may have installation issues)
# textract>=1.6.0

# System Dependencies (install via brew/package manager):
# For legacy .doc file support:
#   brew install antiword
# 
# Alternative tools (optional):
#   catdoc - not available via brew, install from source if needed
#
# Installation commands:
#   macOS: brew install antiword
#   Ubuntu/Debian: sudo apt-get install antiword
#   CentOS/RHEL: sudo yum install antiword
#
# Verify installation:
#   which antiword  # should return path to antiword executable
#
# Test .doc file processing:
#   python3 -c "from enhanced_text_extractor import EnhancedTextExtractor; print('antiword available:', EnhancedTextExtractor().extract_text_with_fallback.__doc__)"
